requirements.txt
cachetools==4.2.4
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.10.2
google-api-python-client==2.31.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
numpy==1.22.1
oauth2client==4.1.3
oauthlib==3.2.2
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
uritemplate==4.1.1
streamlit